Split get_probability_no_exceedance in 2 functions
